+++ mplapack/reference/Clalsa.cpp
@@ -227,7 +227,7 @@ void Clalsa(INTEGER const icompq, INTEGER const smlsiz, INTEGER const n, INTEGER
     //     Finally go through the left singular vector matrices of all
     //     the other subproblems bottom-up on the tree.
     //
-    j = pow(2, nlvl);
+    j = pow((double)2, (double)nlvl);
     sqre = 0;
     //
     for (lvl = nlvl; lvl >= 1; lvl = lvl - 1) {
@@ -240,7 +240,7 @@ void Clalsa(INTEGER const icompq, INTEGER const smlsiz, INTEGER const n, INTEGER
             lf = 1;
             ll = 1;
         } else {
-            lf = pow(2, (lvl - 1));
+            lf = (INTEGER)pow((double)2, (double)(lvl - 1));
             ll = 2 * lf - 1;
         }
         for (i = lf; i <= ll; i = i + 1) {
@@ -274,7 +274,7 @@ statement_170:
             lf = 1;
             ll = 1;
         } else {
-            lf = pow(2, (lvl - 1));
+            lf = (INTEGER)pow((double)2, (double)(lvl - 1));
             ll = 2 * lf - 1;
         }
         for (i = ll; i >= lf; i = i - 1) {
