+++ mplapack/reference/Rlaed0.cpp
@@ -150,10 +150,10 @@ statement_10:
         //
         temp = log(castREAL(n)) / log(two);
         lgn = castINTEGER(temp);
-        if (pow(2, lgn) < n) {
+        if ((INTEGER)pow((double)2, (double)lgn) < n) {
             lgn++;
         }
-        if (pow(2, lgn) < n) {
+        if ((INTEGER)pow((double)2, (double)lgn) < n) {
             lgn++;
         }
         iprmpt = indxq + n + 1;
