+++ mplapack/reference/Rlasda.cpp
@@ -179,7 +179,7 @@ void Rlasda(INTEGER const icompq, INTEGER const smlsiz, INTEGER const n, INTEGER
     //
     //     Now conquer each subproblem bottom-up.
     //
-    j = pow(2, nlvl);
+    j = (INTEGER)pow((double)2, (double)nlvl);
     INTEGER lvl = 0;
     INTEGER lvl2 = 0;
     INTEGER lf = 0;
@@ -197,7 +197,7 @@ void Rlasda(INTEGER const icompq, INTEGER const smlsiz, INTEGER const n, INTEGER
             lf = 1;
             ll = 1;
         } else {
-            lf = pow(2, (lvl - 1));
+            lf = (INTEGER)pow((double)2, (double)(lvl - 1));
             ll = 2 * lf - 1;
         }
         for (i = lf; i <= ll; i = i + 1) {
