+++ mplapack/reference/Rstedc.cpp
@@ -111,10 +111,10 @@ void Rstedc(const char *compz, INTEGER const n, REAL *d, REAL *e, REAL *z, INTEG
             lwmin = 2 * (n - 1);
         } else {
             lgn = castINTEGER(log(castREAL(n)) / log(two));
-            if (pow(2, lgn) < n) {
+            if ((INTEGER)pow((double)2, (double)lgn) < n) {
                 lgn++;
             }
-            if (pow(2, lgn) < n) {
+            if ((INTEGER)pow((double)2, (double)lgn) < n) {
                 lgn++;
             }
             if (icompz == 1) {
