+++ mplapack/test/lin/common/Cchkgt.cpp
@@ -110,7 +110,7 @@ void Cchkgt(bool *dotype, INTEGER const nn, INTEGER *nval, INTEGER const nns, IN
         //        Do for each value of N in NVAL.
         //
         n = nval[in - 1];
-        m = max(n - 1, 0);
+        m = max(n - 1, (INTEGER)0);
         lda = max((INTEGER)1, n);
         nimat = ntypes;
         if (n <= 0) {
