+++ mplapack/test/lin/common/Cchklqtp.cpp
@@ -108,7 +108,7 @@ void Cchklqtp(REAL const thresh, bool const tsterr, INTEGER const nm, INTEGER *m
             //           Do for each value of L
             //
             minmn = min(m, n);
-            for (l = 0; l <= minmn; l = l + max(minmn, 1)) {
+            for (l = 0; l <= minmn; l = l + max(minmn, (INTEGER)1)) {
                 //
                 //              Do for each possible value of NB
                 //
