+++ mplapack/test/lin/common/Cchkpt.cpp
@@ -322,7 +322,7 @@ void Cchkpt(bool *dotype, INTEGER const nn, INTEGER *nval, INTEGER const nns, IN
                 Cpttrs("Lower", n, 1, &d[(n + 1) - 1], &e[(n + 1) - 1], x, lda, info);
                 ainvnm = max({ainvnm, RCasum(n, x, 1)});
             }
-            rcondc = one / max(one, anorm * ainvnm);
+            rcondc = one / max(one, REAL(anorm * ainvnm));
             //
             for (irhs = 1; irhs <= nns; irhs = irhs + 1) {
                 nrhs = nsval[irhs - 1];
