+++ mplapack/test/lin/common/Cdrvhe_rk.cpp
@@ -163,7 +163,7 @@ void Cdrvhe_rk(bool *dotype, INTEGER const nn, INTEGER *nval, INTEGER const nrhs
     //
     for (in = 1; in <= nn; in = in + 1) {
         n = nval[in - 1];
-        lda = max(n, 1);
+        lda = max(n, (INTEGER)1);
         xtype = 'N';
         nimat = ntypes;
         if (n <= 0) {
