+++ mplapack/test/lin/common/Cgbt02.cpp
@@ -61,7 +61,7 @@ void Cgbt02(const char *trans, INTEGER const m, INTEGER const n, INTEGER const k
     //     ..
     //     .. Executable Statements ..
     //
-    //     Quick return if N = 0 pr NRHS = 0
+    //     Quick return if N = 0 or NRHS = 0
     //
     const REAL zero = 0.0;
     if (m <= 0 || n <= 0 || nrhs <= 0) {
@@ -78,7 +78,7 @@ void Cgbt02(const char *trans, INTEGER const m, INTEGER const n, INTEGER const k
     INTEGER i1 = 0;
     INTEGER i2 = 0;
     for (j = 1; j <= n; j = j + 1) {
-        i1 = max(kd + 1 - j, 1);
+        i1 = max(kd + 1 - j, (INTEGER)1);
         i2 = min(kd + m - j, kl + kd);
         anorm = max({anorm, RCasum(i2 - i1 + 1, &a[(i1 - 1) + (j - 1) * lda], 1)});
     }
@@ -114,7 +114,7 @@ void Cgbt02(const char *trans, INTEGER const m, INTEGER const n, INTEGER const k
         if (xnorm <= zero) {
             resid = one / eps;
         } else {
-            resid = max(resid, ((bnorm / anorm) / xnorm) / eps);
+            resid = max(resid, REAL(((bnorm / anorm) / xnorm) / eps));
         }
     }
     //
