+++ mplapack/test/lin/common/Cpot06.cpp
@@ -107,7 +107,7 @@ void Cpot06(const char *uplo, INTEGER const n, INTEGER const nrhs, COMPLEX *a, I
         if (xnorm <= zero) {
             resid = one / eps;
         } else {
-            resid = max(resid, ((bnorm / anorm) / xnorm) / eps);
+            resid = max(resid, REAL(((bnorm / anorm) / xnorm) / eps));
         }
     }
     //
