+++ mplapack/test/lin/common/Cppt02.cpp
@@ -99,7 +99,7 @@ void Cppt02(const char *uplo, INTEGER const n, INTEGER const nrhs, COMPLEX *a, C
         if (xnorm <= zero) {
             resid = one / eps;
         } else {
-            resid = max(resid, ((bnorm / anorm) / xnorm) / eps);
+            resid = max(resid, REAL(((bnorm / anorm) / xnorm) / eps));
         }
     }
     //
